Prevent DefaultFileStore from throwing a null pointer exception upon file removal if neither exists, the file to delete and the directory it's assumed to be.

+++ common/src/org/riotfamily/common/web/file/DefaultFileStore.java
@@ -228,7 +228,9 @@ public class DefaultFileStore implements FileStore, ServletContextAware,
 		File file = retrieve(uri);
 		file.delete();
 		File dir = file.getParentFile();
-		if (dir.list().length == 0 && !dir.equals(baseDir)) {
+		if (dir.isDirectory() && dir.list().length == 0 &&
+			!dir.equals(baseDir)) {
+			
 			dir.delete();
 		}
 	}
